Updating trial_path to use main value
The trial_path output wasn't being used and was mannualy filled to always
get from the /1 folder. This patch changes it to use the main value and get
the last execution correctly. Avoiding errors when the execution is not
completed in the first execution and always copying the latest model on
the subsequent executions.